+++ test/test_metalearners.jl
@@ -48,6 +48,7 @@ x_learner_binary = XLearner(x, t, Float64.([rand() < 0.4 for i in 1:100]))
 estimate_causal_effect!(x_learner_binary)
 
 rlearner = RLearner(x, t, y)
+residuals = CausalELM.weight_trick(rlearner, t, y)
 estimate_causal_effect!(rlearner)
 
 # Testing initialization with DataFrames
@@ -57,8 +58,8 @@ r_learner_df = RLearner(x_df, t_df, y_df)
 dr_learner = DoublyRobustLearner(x, t, y)
 X, T, Y = CausalELM.generate_folds(
     dr_learner.X, dr_learner.T, dr_learner.Y, dr_learner.folds
-    )
-τ̂ = CausalELM.doubly_robust_formula!(dr_learner, X, T, Y)
+)
+causal_effect, marginal_effect = CausalELM.doubly_robust_formula!(dr_learner, X, T, Y)
 estimate_causal_effect!(dr_learner)
 
 # Testing Doubly Robust Estimation with a binary outcome
@@ -82,7 +83,9 @@ estimate_causal_effect!(dr_learner_df)
 
     @testset "S-Learner Estimation" begin
         @test isa(slearner1.causal_effect, Array{Float64})
+        @test all(isnan, slearner1.marginal_effect) == false
         @test isa(s_learner_binary.causal_effect, Array{Float64})
+        @test all(isnan, s_learner_binary.marginal_effect) == false
     end
 end
 
@@ -98,7 +101,9 @@ end
 
     @testset "T-Learner Estimation" begin
         @test isa(tlearner1.causal_effect, Array{Float64})
+        @test all(isnan, tlearner1.marginal_effect) == false
         @test isa(t_learner_binary.causal_effect, Array{Float64})
+        @test all(isnan, t_learner_binary.marginal_effect) == false
     end
 end
 
@@ -133,7 +138,9 @@ end
         @test typeof(xlearner3.μ₁) <: CausalELM.ELMEnsemble
         @test xlearner3.ps isa Array{Float64}
         @test xlearner3.causal_effect isa Array{Float64}
+        @test all(isnan, xlearner3.marginal_effect) == false
         @test x_learner_binary.causal_effect isa Array{Float64}
+        @test all(isnan, x_learner_binary.marginal_effect) == false
     end
 end
 
@@ -149,9 +156,11 @@ end
 
     @testset "R-learner estimation" begin
         @test rlearner.causal_effect isa Vector
+        @test length(residuals) == length(y)
         @test length(rlearner.causal_effect) == length(y)
         @test eltype(rlearner.causal_effect) == Float64
         @test all(isnan, rlearner.causal_effect) == false
+        @test all(isnan, rlearner.marginal_effect) == false
     end
 end
 
@@ -167,7 +176,8 @@ end
     end
 
     @testset "Calling estimate_effect!" begin
-        @test length(τ̂) === length(dr_learner.Y)
+        @test length(causal_effect) === length(dr_learner.Y)
+        @test length(marginal_effect) === length(dr_learner.Y)
     end
 
     @testset "Doubly Robust Learner Estimation" begin
@@ -175,6 +185,7 @@ end
         @test length(dr_learner.causal_effect) === length(y)
         @test eltype(dr_learner.causal_effect) == Float64
         @test all(isnan, dr_learner.causal_effect) == false
+        @test all(isnan, dr_learner.marginal_effect) == false
         @test dr_learner_df.causal_effect isa Vector
         @test length(dr_learner_df.causal_effect) === length(y)
         @test eltype(dr_learner_df.causal_effect) == Float64